Backend/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.1.2
django-cors-headers==4.7.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
frozenlist==1.5.0
idna==3.7
multidict==6.1.0
php==1.2.1
pillow==11.0.0
propcache==0.3.0
psycopg2-binary==2.9.10
pyasn1==0.6.0
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
sqlparse==0.5.1
twilio==9.4.6
tzdata==2024.2
urllib3==2.2.2
yarl==1.18.3
gunicorn==20.1.0  
whitenoise
reportlab
django-redis==4.12.1


# weasyprint
# cffi
# cairocffi

# FastAPI (v1 API layer — vendor onboarding & new sprint endpoints)
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
pydantic[email]==2.9.2
python-multipart==0.0.12
